Major rewrite to use better Javascript.
I'm a bad person. I refactored a ton of things at once.

Buffer was replaced by ByteArray. It is a rewrite with a different
purpose, essentially being a uint8_t[], but with helpful accessors. It
also fully supports UTF-8, which I think is new.

Long was rewritten. It's basically the same, but much cleaner, and
shifts are definitely more efficient.

I have tests for Serpent/Tiger/Whirlpool, so I know that those work, as
do most of ByteArray and Long. I do not know how well Crypt, Hmac, and
pbkdf2 work.